Updated library version to v8.45
This commits updates the library version to v8.45 and changes the
build process of the library to not generate files dynamically via
configure. Instead, the library comes with all dynamically created
files. This prevents name clashes for example with the config.h
and allows for better control of build parameters.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>